Split the Run module.
